set state added

+++ app/src/main/java/com/crewmates/autolibodb/MainActivity.kt
@@ -50,7 +50,7 @@ class MainActivity : FragmentActivity(), OnMapReadyCallback {
          val fuel = intent.getIntExtra("fuel", 0)
          val temp = intent.getIntExtra("temperature", 0)
          val idRental = intent.getIntExtra("idRental", 0)
-         val oilChange = intent.getDoubleExtra("nextOilChange", 0.0)
+         val oilChange = intent.getIntExtra("nextOilChange", 0)
          Prefs.idRental = idRental
          Prefs.oilChange = oilChange
          Prefs.temperature= temp
